Name function update
Added conditional for no-input entries. Has a bug where using getchar() while recursively looping the program creates issues after the first use of the warning entering input for user without a name. The following tries make it so that user has to press enter twice and also the function then removes the first letter from the name given after that. It is currently being worked around by just using return 1 instead of the error checking function which kicks the user back to the main menu. Am currently working on this, will try to use an error-detecting loop to first check and then move on to the function.