Gender Select added
Cleaned up code a bit in the file. Added a Gender selection module. Works good for a scan f, added printf functions to not throw off the puts function outputs.
Need to add a name filtering function and possibly a class selection? Now the main character struct needs to be created along with the rooms.